+++ Sources/KeyboardKit/Previews/KeyboardLayoutProvider+Preview.swift
@@ -22,8 +22,8 @@ public extension KeyboardLayoutProvider where Self == PreviewKeyboardLayoutProvi
  */
 public class PreviewKeyboardLayoutProvider: KeyboardLayoutProvider {
     
-    public init(context: KeyboardContext = .preview) {
-        let inputProvider = StandardInputSetProvider(context: context)
+    public init(keyboardContext: KeyboardContext = .preview) {
+        let inputProvider = StandardInputSetProvider(keyboardContext: keyboardContext)
         provider = StandardKeyboardLayoutProvider(
             keyboardContext: .preview,
             inputSetProvider: inputProvider)